+++ semana9/labex/src/Components/Header.js
@@ -1,5 +1,6 @@
 import styled from 'styled-components';
 import logo from '../img/x.png';
+import ContactPhoneIcon from '@material-ui/icons/ContactPhone';
 /*
 @media (max-width: 800px) {
     flex-wrap: wrap;
@@ -10,7 +11,6 @@ const Principal = styled.div`
     display:flex;
     flex:1;
     align-items:center;
-    font-size:50px;
     background: linear-gradient(180deg, #FFFFFF 0%, rgba(255, 255, 255, 0) 100%), #000000;
     color:#ffffff;
     justify-content:space-between;
@@ -19,21 +19,34 @@ const Logo = styled.div`
     display:flex;
     align-items:center;
     margin-left:4px;
+    font-size:50px;
 `
 const ImagemLogo = styled.img`
     height:100px;
     background-color:#ffffff;
     border:2px solid black;
 `
+const IconeAdmin = styled.div`
+    display:flex;
+    flex-direction:column;
+    align-items:center;
+    font-size:20px;
+    margin: 0 30px;
+`
 
-export default function Header(){
-    return(
+export default function Header() {
+   
+        
+    return (
         <Principal>
             <Logo>
                 <ImagemLogo src={logo} />
                 <h2>LABEX</h2>
             </Logo>
-            <div>Area Do Admin</div>
+            <IconeAdmin>
+                <ContactPhoneIcon style={{ fontSize: 60 }}/>
+                Central de atendimento: (99) 9999-9999
+            </IconeAdmin>
         </Principal>
     )
 }
