+++ semana9/labex/src/Pages/homepage/Style.js
@@ -1,0 +1,74 @@
+import styled from 'styled-components'
+
+export const Container = styled.div`
+    display:flex;
+    flex-direction:column;
+`
+export const TituloContainer = styled.h2`
+    font-family:Apple Chancery, cursive;
+    text-transform:uppercase;
+    font-size:20px;
+    display:flex;
+    justify-content:flex-start;
+    margin:0 10px;
+`
+export const ListagemViagem = styled.div`
+    margin: 10px 50px;
+    width:90vw;
+    min-height:50px;
+    min-width:100px;
+    display:flex;
+    justify-content:space-between;
+    align-items:center;
+`
+export const ExplorePlanetas = styled.div`
+    margin: 10px 50px;
+    width:90vw;
+    min-height:100px;
+    min-width:100px;
+    background-color:#C4C4C4;
+    justify-content:space-between;
+`
+export const Inspiracoes = styled.div`
+    margin: 10px 50px;
+    width:90vw;
+    min-height:100px;
+    min-width:100px;
+    background-color:#C4C4C4;
+`
+
+export const ImagemPlaneta = styled.img`
+    height:250px;
+    width:300px;
+    margin:10px;
+    border:3px solid white;
+`
+export const ImagemPlanetaViagens = styled.img`
+    height:350px;
+    width:340px;
+    margin:10px;
+    border:3px solid white;
+`
+export const Planetas = styled.div`
+    display:flex;
+    min-height:100px;
+    min-width:100px;
+    align-items:center;
+    justify-content:space-around;
+`
+export const PlanetasFilho = styled.div`
+    display:flex;
+    flex-direction:column;
+    align-items:center;
+`
+export const PlanetasFilhoInspiracoes = styled.div`
+    display:flex;
+    flex-direction:column;
+    align-items:center;
+    justify-content:center;
+`
+export const Paragrafo = styled.p`
+    font-weight:bold;
+    font-size:20px;
+    text-align:center;
+`
